ExplicitConnectivity now uses ArrayHandle.allocate + std::copy in FillViaCopy.
The original implementation use the naive approach of the device adapter
to copy the memory to the dynamic array handle. The issue with this approach
is that it would do a upload to device, and than a copy on device. Instead
we should prefer a copy on host, and than an upload to device.